Move event filters to separate module
